pyproject.toml
[tool.poetry]
name = "transaction_analyzer"
version = "0.1.0"
description = "Application for transaction analysis"
authors = ["Ianioglo Vladimir <skypromd@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0"
python-dotenv = "^1.0"
requests = "^2.28"
pytest = "^7.0"
pytest-mock = "^3.10"
black = "24.3.0"
isort = "^5.12"
flake8 = "^6.0"
mypy = "^1.0"
openpyxl = "^3.1"
python-dateutil = "^2.8"

[tool.poetry.group.dev.dependencies]
types-requests = "^2.32.0.20250306"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
)/
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv/"
explicit_package_bases = true
namespace_packages = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"